+++ portal/blogs/blog_check_strategy.py
